feat: Add the infrastructure for a navbar
Addition of the navbar infrastructure. Unfortunately, we can't really
add more to this without adding more screens.
